==> Azure-Model-Deployment/4_API.ipynb <==
import os
import json
import requests
from dotenv import load_dotenv

load_dotenv() 
scoring_uri = os.getenv("AZURE_SCORING_URI")
api_key     = os.getenv("AZURE_API_KEY")

headers = {
    "Content-Type": "application/json",
    "Authorization": f"Bearer {api_key}"
}

payload = json.dumps([{
    "lag_1": 0.306493,
    "lag_7": 0.249184,
    "lag_14": 0.112441,
    "rolling_mean_7": 0.154945,
    "dow_sin": 0.00,
    "dow_cos": 0.356896,
    "month_sin": 0.50,
    "month_cos": 1.0,
    "temporada_alta": 0,
    "es_festivo": 0,
    "es_viernes_o_sabado": 1,
    "es_pandemia": 0
}])

try:
    response = requests.post(scoring_uri, headers=headers, data=payload, timeout=30)
    response.raise_for_status()
    print("Respuesta endpoint:", response.json())
except requests.exceptions.RequestException as e:
    print("Error al invocar el endpoint:", e)
